Trickle ICE on client side (Java API and demo)
The candidates recvd from server are stored in a blocking
queue which can be read using the Java API. The client app
is responsible of sending the candidate to the corresponding
sink in the appropiate tree (treeId and sinkId are also included).

Using the Java API, the client app can notify the server of a
new gathered ICE candidate from a browser instance.

Change-Id: I278aeefa45ad6109f8e49502ac01c6d65a479485